Removed explicit params

+++ Source/UnitTests/AssetManagerUnitTests.cpp
@@ -36,7 +36,7 @@ UNIT_TEST(AssetManagerUnitTests, MultiThreadedAssetLoadingUnloading)
 	std::vector<std::future<UnitTest::Result>> loadResults{};
 	std::vector<std::future<void>> unloadResults{};
 
-	for (int i = 0; i < ThreadPool::Get().NumberOfThreads(); i += 2)
+	for (size_t i = 0; i < ThreadPool::Get().NumberOfThreads(); i += 2)
 	{
 		loadResults.emplace_back(ThreadPool::Get().Enqueue(loadAssets));
 		unloadResults.emplace_back(ThreadPool::Get().Enqueue(unloadAssets));
